refactor(discord): extract watcher turn stream collector
Move the watcher turn stream collector into tmux_watcher/turn_stream_collector.rs. Start anchor: initial decoded buffer seed (decoded_data = utf8_decoder.decode(...)); end anchor: active read/parse loop through update_streaming_status_tick fallthrough and active-read tail capture.\n\nControl-flow mapping: escaping returns=0, escaping breaks=0, escaping continues=3. The three outer watcher-loop continues map 1:1 to CollectOutcome::ContinueWatcherLoop at the callsite; inner read-loop breaks/continues remain internal. Fallthrough maps to CollectOutcome::Fallthrough.\n\nThreading: 65 distinct locals cross the collector boundary (15 persistent copyback/direct state items + 50 collected fallthrough/tail fields with ActiveReadState expanded). commit_collect_state!/commit_persistent_collect_state! run before every collector return.\n\nMaintenance gates: tmux_watcher.rs ratcheted 5975 -> 5218; new turn_stream_collector.rs is 1158 lines with per-file cap admission and giant-file registry entry. Ran python3 scripts/generate_inventory_docs.py, python3 scripts/check_agent_maintenance_docs.py (agent-maintenance freshness check passed), bash scripts/ci-script-checks.sh (exit 0), and cargo fmt.
