+++ ARCHITECTURE.md
@@ -650,6 +650,7 @@ src/
 │   │   │   ├── tests.rs
 │   │   │   ├── turn_identity.rs
 │   │   │   ├── turn_identity_tests.rs
+│   │   │   ├── turn_stream_collector.rs
 │   │   │   ├── two_message_panel.rs
 │   │   │   ├── utf8_chunk_decoder.rs
 │   │   │   └── utf8_chunk_decoder_tests.rs
