+++ scripts/audit_maintainability_config.toml
@@ -28,6 +28,7 @@
 "src/services/discord/prompt_builder/**" = 700
 "src/services/discord/runtime_bootstrap/**" = 700
 "src/services/discord/tmux_watcher/streaming_status_tick.rs" = 963 # 2026-07-08 #4229 S3: approved contiguous verbatim extraction of the throttled streaming status tick (orphan reclaim/suppression/panel create-bind/rollover/re-anchor/placeholder edit); below giant threshold, but irreducible as one review-verified move.
+"src/services/discord/tmux_watcher/turn_stream_collector.rs" = 1158 # 2026-07-08 #4229 S4: approved contiguous verbatim extraction of the turn stream collector (seed restore/first parse/forward/monitor auto-turn claim/active read-parse loop); below giant threshold, but irreducible as one review-verified move.
 "src/services/discord/tmux_watcher/**" = 700
 "src/services/discord/turn_bridge/headless_delivery.rs" = 700
 "src/services/discord/turn_bridge/status_panel.rs" = 700
